requirements.txt
asgiref==3.7.2
Django==4.2.10
django-autoslug==1.9.9
django-colorfield==0.11.0
django-grappelli==3.0.8
django-jazzmin==2.6.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-slugify==8.0.3
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
webcolors==1.13
